Made the amount of time between ticks a fixed constant. Reduced it so appears more like a continuous system.

+++ app/src/main/java/com/oreilly/demo/android/pa/uidemo/view/GameView.java
@@ -63,7 +63,7 @@ public final class GameView extends View {
     {
         TextView view = (TextView)((DefaultClockModel)(monsterGame.getClockModel())).getScore_view();
         view.setText("Current Score: " + new Integer(value).toString()  + "\n" +
-        "Time Remaining: " + new Integer(monsterGame.getClockModel().get_time_to_wait()).toString());
+        "Time Remaining: " + new Float(monsterGame.getClockModel().get_time_to_wait()).toString());
     }
 
     @Override
